-too many changes, can't remember details, see the diff. Anyway nobody is reading this :-(

+++ customroot/usr/www/cgi-bin/packages_ipkg.cgi
@@ -9,7 +9,6 @@ write_header "Alt-F Package Manager"
 s="<strong>"
 es="</strong>"
 
-#if ! test -f /usr/bin/ipkg-cl; then
 if ! ipkg status >/dev/null; then
 
 	disks=$(ls /dev/sd?) >/dev/null 2>&1
@@ -30,19 +29,34 @@ if ! ipkg status >/dev/null; then
 
 else
 
-ipkg-cl -V0 info | awk '
-	/Package:/{i++; nm=$2; pkg[i] = nm} # this relies on Package being the first field 
-	/Version:/{ver[i] = $2} 
-	/Source:/{url[i] = $2} 
-	/Description:/{des[i] = substr($0, index($0,$2))} 
-	/Status:/{if ($4 == "installed") inst[nm] = i; else uinst[nm] = i} 
-	END {
-		print "<form action=\"/cgi-bin/packages_ipkg_proc.cgi\" method=post> \
-		<fieldset><legend><strong> Installed Packages </strong></legend> \
-		<table><tr> \
-			<th>Name</th><th>Version</th><th></th><th></th><th>Description</th> \
-		</tr>"
+feed=$(awk '/^src /{print $3}' /etc/ipkg.conf)
 
+ipkg-cl -V0 info | awk -v confeeds="$feed" '
+	BEGIN {
+		nfeeds = split(confeeds, feeds);
+	}
+	/Package:/ { i++; nm=$2; pkg[i] = nm } # this relies on Package being the first field 
+	/Version:/ { ver[i] = $2 } 
+	/Source:/ { url[i] = $2 } 
+	/Description:/ { des[i] = substr($0, index($0,$2)) } 
+	/Status:/ { if ($4 == "installed")
+			inst[nm] = i;
+		else { uinst[nm] = i; ucnt++;}
+	}  
+	END {
+		printf "<form action=\"/cgi-bin/packages_ipkg_proc.cgi\" method=post> \
+			<fieldset><legend><strong> Configure Feeds </strong></legend> \
+			<input type=hidden name=nfeeds value=%d>", nfeeds
+		for (i=1; i<=nfeeds; i++) 
+			printf "Feed %d: <input type=text size=50 name=feed_%d value=%s><br>", i, i, feeds[i];
+		print "<input type=submit name=changeFeed value=ChangeFeed> \
+			<input type=submit name=defaultFeed value=DefaultFeed><br><br> \
+			Update package list: <input type=submit name=updatelist value=UpdateList> \
+			</fieldset><br> \
+			<fieldset><legend><strong> Installed Packages </strong></legend> \
+			<table><tr> \
+				<th>Name</th><th>Version</th><th></th><th></th><th>Description</th> \
+			</tr>"
 
 		update = 0;
 		for (i=1; pkg[i] != ""; i++) {
@@ -51,7 +65,7 @@ ipkg-cl -V0 info | awk '
 				if (nm in uinst) {	# new version available, old has missing info
 					j = uinst[nm]; v = ver[inst[nm]]; update++;
 					upd=sprintf("<td><input type=submit name=%s value=Update></td>", nm);
-					delete uinst[nm]; delete inst[nm];			
+					delete uinst[nm]; ucnt--; delete inst[nm];
 				} else {
 					j = i; v = ver[i];
 					upd="<td></td>";
@@ -64,29 +78,40 @@ ipkg-cl -V0 info | awk '
 			}
 		}
 	
+		print "<tr><td><br></td></tr>"
 		if (update != 0)
-			print "<tr><td></td><td></td><td></td> \
+			print "<tr><td colspan=2>UpdateAll</td> \
 				<td><input type=submit name=updateall value=UpdateAll></td></tr>"
 
-		print "</table></fieldset> \
-			<br><fieldset><legend><strong> Available Packages </strong></legend> \
-			<table><tr> \
-			<th>Name</th><th>Version</th> \
-			<th></th><th>Description</th></tr>"
+		print "<script type=\"text/javascript\"> \
+				function ask() { \
+					return confirm(\"All files and configurations will be erased. You will have to reinstall ipkg on a disk partition.\"); \
+				} \
+			</script> \
+			<tr><td colspan=2>Remove all installed</td> \
+				<td><input type=submit name=removeall value=RemoveAll onclick=\"return ask()\"></td></tr> \
+			</table></fieldset> \
+			<br><fieldset><legend><strong> Available Packages </strong></legend><table>"
 
-		for (i=1; pkg[i] != ""; i++) {
-			nm = pkg[i];
-			if (nm in uinst) {
-				printf "<tr><td><a href=\"%s\">%s</a></td><td>%s</td>",
+		if (ucnt == 0) {
+			print "None"
+		} else {
+			print "<tr><th>Name</th><th>Version</th> \
+				<th></th><th>Description</th></tr>"
+
+			for (i=1; pkg[i] != ""; i++) {
+				nm = pkg[i];
+				if (nm in uinst) {
+					printf "<tr><td><a href=\"%s\">%s</a></td><td>%s</td>",
 					url[i], nm, ver[i];
-				printf "<td><input type=submit name=%s value=Install></td>", nm;
-				printf "<td>%s</td></tr>\n\n", des[i];
+					printf "<td><input type=submit name=%s value=Install></td>", nm;
+					printf "<td>%s</td></tr>\n\n", des[i];
+				}
 			}
 		}
-		print "</table></fieldset><br> \
-		<input type=submit name=updatelist value=UpdateList> \
-		<input type=submit name=configfeed value=ConfigureFeed> \
-		</form>"
+		printf ("<tr><td><br></td></tr><tr><td colspan=2>Install all available</td> \
+			<td><input type=submit name=installall value=InstallAll></td></tr>\
+		</table></fieldset></form>", pkgfeed)
 	}'
 
 fi
